Switch to 16-bit control word

+++ arduino/pincontrol/shiftctrl.h
@@ -7,16 +7,16 @@
 class ShiftCtrl
 {
         ShiftOutExt shext;
-        uint8_t init;
+        uint16_t init;
     public:
-        ShiftCtrl(uint8_t init);
+        ShiftCtrl(uint16_t init);
         void setup();
         void commit();
         void reset();
         void set(uint8_t pin);
         void clear(uint8_t pin);
     private:
-        uint8_t buffer;
+        uint16_t buffer;
 };
 
 #endif /* SHIFT_CTRL_H */
